restructure and initial chease setup

+++ setup_test_files.bash
@@ -1,11 +1,12 @@
 #!/bin/bash
 
+# Should be run from PDS repo directory
+
 set -euo pipefail
 
-given_dir="pds/ymmsl_files"
+given_dir="ymmsl_files"
 
 for file in "$given_dir"/.*.ymmsl; do
-  echo "Processing: $file"
   # Skip . and ..
   [[ "$file" != "$given_dir"/.*.ymmsl ]] && continue
 
@@ -15,6 +16,12 @@ for file in "$given_dir"/.*.ymmsl; do
   # Remove the leading dot
   new_filename="${filename#.}"
 
+  # Skip if already exists
+  if [ -f "$given_dir/$new_filename" ]; then
+    continue
+  fi
+
+  echo "Processing: $file"
   # Use sed to replace the matching substrings
   sed "s|\[PWD_PLACEHOLDER\]|$(pwd)|g"  "$file" > "$given_dir/$new_filename"
 done
